test(bash): isolate TestBashTimeout from -race goroutine starvation
CI's `go test -race ./...` flaked TestBashTimeout with elapsed=5.001s
(full sleep duration) despite Timeout=100ms. Production code is
correct — Bash.Execute uses exec.CommandContext, which spawns an
internal goroutine that calls Process.Kill() when the runCtx
deadline fires. Under the race detector, that goroutine gets
starved long enough that the kill arrives only after `sleep 5`
exits naturally, defeating the assertion.

The race detector is the right tool for finding data races in our
Go code, not for validating goroutine-scheduling-sensitive timing.
The conventional fix is the `!race` build tag — TestBashTimeout
runs under `go test ./...` (where the kill goroutine schedules
promptly), and is excluded under `go test -race ./...`.

All other Bash tests (refusals, output truncation, env scrubbing,
caller-timeout cap) remain in bash_test.go and run under both
modes — they're functional, not timing-sensitive.

Verified:
- go test ./internal/tools/builtin/... — TestBashTimeout passes (0.10s)
- go test -race ./internal/tools/builtin/... — TestBashTimeout excluded
- go test ./... and go test -race ./... — full suite green
